+++ lib/src/specs/typedef.dart
@@ -9,14 +9,16 @@ import 'package:meta/meta.dart';
 import '../base.dart';
 import '../mixins/annotations.dart';
 import '../mixins/dartdoc.dart';
+import '../mixins/generics.dart';
 import '../visitors.dart';
 import 'expression.dart';
+import 'reference.dart';
 
 part 'typedef.g.dart';
 
 @immutable
 abstract class TypeDef extends Object
-    with HasAnnotations, HasDartDocs
+    with HasAnnotations, HasDartDocs, HasGenerics
     implements Built<TypeDef, TypeDefBuilder>, Spec {
   factory TypeDef([void Function(TypeDefBuilder)? updates]) = _$TypeDef;
 
@@ -39,7 +41,7 @@ abstract class TypeDef extends Object
 }
 
 abstract class TypeDefBuilder extends Object
-    with HasAnnotationsBuilder, HasDartDocsBuilder
+    with HasAnnotationsBuilder, HasDartDocsBuilder, HasGenericsBuilder
     implements Builder<TypeDef, TypeDefBuilder> {
   factory TypeDefBuilder() = _$TypeDefBuilder;
 
@@ -50,6 +52,9 @@ abstract class TypeDefBuilder extends Object
 
   @override
   ListBuilder<String> docs = ListBuilder<String>();
+
+  @override
+  ListBuilder<Reference> types = ListBuilder<Reference>();
 
   String? name;